requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
sqlalchemy
psycopg2-binary
passlib[argon2]
python-jose[cryptography]
python-dotenv
boto3
firebase-admin
requests